delete on exit only if really needed

+++ tika-core/src/main/java/org/apache/tika/fork/ForkClient.java
@@ -323,8 +323,9 @@ class ForkClient {
             }
         }
         if (jar != null) {
-            jar.delete();
-            jar.deleteOnExit();
+            if (!jar.delete()) {
+                jar.deleteOnExit();
+            }
         }
     }
